Extract About Titlebar Debug subsystem from cmuxApp.swift into CmuxDebugWindowsUI
Pull the self-contained About-titlebar debug cluster out of the cmuxApp.swift
god file into a new domain UI package, CmuxDebugWindowsUI, as a Coordinator plus
value types, store, window controller, and SwiftUI view.

What moved (cmuxApp.swift -> Packages/CmuxDebugWindowsUI):
- AboutWindowKind, TitlebarVisibilityOption, TitlebarToolbarStyleOption value enums
- AboutTitlebarDebugOptions value type
- AboutTitlebarDebugStore (was a singleton ObservableObject -> @MainActor @Observable)
- AboutTitlebarDebugWindowController (NSWindowController)
- AboutTitlebarDebugView (SwiftUI)
New: DebugWindowsCoordinator (@MainActor @Observable) owns the store + lazily owns
the window controller; WindowDecorating protocol seam.

Seams (constructor-injected, god-type reaches inverted):
- AppDelegate.shared?.applyWindowDecorations(to:) -> WindowDecorating protocol.
  AppDelegate conforms directly (it already has the matching method) and is
  injected into DebugWindowsCoordinator at the composition root, held WEAKLY by
  the coordinator/store to avoid a retain cycle (AppDelegate is the singleton
  owner of the coordinator).
- The singleton .shared store/controller are replaced by app-owned instances on
  AppDelegate (debugWindowsCoordinator / aboutTitlebarDebugStore).

Byte-identical behavior:
- Window identifiers, titles, style-mask bits, toolbar identifiers, min/max
  sizes, copy-config payload text, and the didSet-driven reapply are preserved
  exactly. Call sites (Debug menu, openAllDebugWindows, DebugWindowControlsView,
  About/Acknowledgments window controllers) are thin forwards to the app-owned
  coordinator/store, so the runtime sequence is unchanged.
- applyToOpenWindows now guards a nil NSApp (no-op only in a headless test
  process; NSApp is always present in the running app), enabling unit tests.

Tests: 7 Testing-framework behavior tests with a fake WindowDecorating
(defaults, enabled/disabled overrides, empty-title fallback, toolbar toggle,
reset, copy-config payload).

Est lines removed from giant: 405 (cmuxApp.swift 4920 -> 4515). Budget reconciled
(cmuxApp ratcheted down; AppDelegate +11 for composition-root wiring).

Left behind (intentional, not a clean leaf): the #if DEBUG controllers still in
cmuxApp.swift (SidebarDebug, MenuBarExtraDebug, BackgroundDebug, StartupAppearance,
SplitButtonLayout, TabBarBackdropLab, BrowserProfilePopover/ImportHint,
FileExplorerStyle, DebugWindowControls, DebugWindowConfigSnapshot) reach into many
app-target settings/catalog/Option types and Workspace/GhosttyApp statics across
multiple files, so extracting them requires multi-file seam inversion beyond this
PR. The AboutTitlebar cluster is the one fully self-contained, non-DEBUG leaf.

NOTE: app build validated by CI (not run locally per refactor policy).
