+++ Packages/CmuxDebugWindowsUI/Sources/CmuxDebugWindowsUI/Views/AboutTitlebarDebugView.swift
@@ -1,0 +1,130 @@
+#if canImport(AppKit)
+
+public import SwiftUI
+
+/// The editor surface for the About Titlebar Debug window.
+///
+/// Every control reads from and writes through an injected
+/// ``AboutTitlebarDebugStore``; the view holds no other state.
+public struct AboutTitlebarDebugView: View {
+    private let store: AboutTitlebarDebugStore
+
+    /// Creates the editor view bound to a store.
+    ///
+    /// - Parameter store: The store whose options the editor reads and mutates.
+    public init(store: AboutTitlebarDebugStore) {
+        self.store = store
+    }
+
+    public var body: some View {
+        ScrollView {
+            VStack(alignment: .leading, spacing: 14) {
+                Text(String(localized: "debug.aboutTitlebarDebug.title", defaultValue: "About Titlebar Debug"))
+                    .font(.headline)
+
+                editor(for: .about)
+
+                GroupBox("Actions") {
+                    HStack(spacing: 10) {
+                        Button("Reset All") {
+                            store.reset(.about)
+                        }
+                        Button("Reapply to Open Windows") {
+                            store.applyToOpenWindows()
+                        }
+                        Button("Copy Config") {
+                            store.copyConfigToPasteboard()
+                        }
+                    }
+                    .frame(maxWidth: .infinity, alignment: .leading)
+                    .padding(.top, 2)
+                }
+
+                Spacer(minLength: 0)
+            }
+            .padding(16)
+            .frame(maxWidth: .infinity, alignment: .topLeading)
+        }
+        .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
+    }
+
+    private func editor(for kind: AboutWindowKind) -> some View {
+        let overridesEnabled = binding(for: kind, keyPath: \.overridesEnabled)
+
+        return GroupBox(kind.displayTitle) {
+            VStack(alignment: .leading, spacing: 10) {
+                Toggle("Enable Debug Overrides", isOn: overridesEnabled)
+
+                Text("When disabled, cmux uses normal default titlebar behavior for this window.")
+                    .font(.caption)
+                    .foregroundColor(.secondary)
+
+                Divider()
+
+                VStack(alignment: .leading, spacing: 10) {
+                    HStack(spacing: 8) {
+                        Text("Window Title")
+                        TextField("", text: binding(for: kind, keyPath: \.windowTitle))
+                    }
+
+                    HStack(spacing: 10) {
+                        Picker("Title Visibility", selection: binding(for: kind, keyPath: \.titleVisibility)) {
+                            ForEach(TitlebarVisibilityOption.allCases) { option in
+                                Text(option.displayTitle).tag(option)
+                            }
+                        }
+                        Picker("Toolbar Style", selection: binding(for: kind, keyPath: \.toolbarStyle)) {
+                            ForEach(TitlebarToolbarStyleOption.allCases) { option in
+                                Text(option.displayTitle).tag(option)
+                            }
+                        }
+                    }
+
+                    Toggle("Show Toolbar", isOn: binding(for: kind, keyPath: \.showToolbar))
+                    Toggle("Transparent Titlebar", isOn: binding(for: kind, keyPath: \.titlebarAppearsTransparent))
+                    Toggle("Movable by Window Background", isOn: binding(for: kind, keyPath: \.movableByWindowBackground))
+
+                    Divider()
+
+                    Text("Style Mask")
+                        .font(.caption)
+                        .foregroundColor(.secondary)
+
+                    Toggle("Titled", isOn: binding(for: kind, keyPath: \.titled))
+                    Toggle("Closable", isOn: binding(for: kind, keyPath: \.closable))
+                    Toggle("Miniaturizable", isOn: binding(for: kind, keyPath: \.miniaturizable))
+                    Toggle("Resizable", isOn: binding(for: kind, keyPath: \.resizable))
+                    Toggle("Full Size Content View", isOn: binding(for: kind, keyPath: \.fullSizeContentView))
+
+                    HStack(spacing: 10) {
+                        Button(String(localized: "debug.aboutTitlebarDebug.resetAbout", defaultValue: "Reset About")) {
+                            store.reset(kind)
+                        }
+                        Button("Apply Now") {
+                            store.applyToOpenWindows(for: kind)
+                        }
+                    }
+                }
+                .disabled(!overridesEnabled.wrappedValue)
+                .opacity(overridesEnabled.wrappedValue ? 1 : 0.75)
+            }
+            .padding(.top, 2)
+        }
+    }
+
+    private func binding<Value>(
+        for kind: AboutWindowKind,
+        keyPath: WritableKeyPath<AboutTitlebarDebugOptions, Value>
+    ) -> Binding<Value> {
+        Binding(
+            get: { store.options(for: kind)[keyPath: keyPath] },
+            set: { newValue in
+                var updated = store.options(for: kind)
+                updated[keyPath: keyPath] = newValue
+                store.update(updated, for: kind)
+            }
+        )
+    }
+}
+
+#endif
